// admin/src/lib/types/model/Circle.ts
export interface Circle {
  type: 'Circle'
  id: number
  slug: string
  release: boolean
  isMainFixed: boolean
  isOnlyDemo: boolean
  isDemoFixed: boolean
  demoPriority: number
  circleType: string
  name: string
  nameKana: string
  shortName: string
  prefixName: string
  description: string
  commonPlaceOfActivity: string
  commonPlaceOfActivityDetail: string
  commonDateOfActivityMonday: boolean
  commonDateOfActivityTuesday: boolean
  commonDateOfActivityWednesday: boolean
  commonDateOfActivityThursday: boolean
  commonDateOfActivityFriday: boolean
  commonDateOfActivitySaturday: boolean
  commonDateOfActivitySunday: boolean
  commonDateOfActivityDetail: string
  isOnlineActivity: boolean
  onlinePlaceOfActivityDetail: string
  onlineDateOfActivityMonday: boolean
  onlineDateOfActivityTuesday: boolean
  onlineDateOfActivityWednesday: boolean
  onlineDateOfActivityThursday: boolean
  onlineDateOfActivityFriday: boolean
  onlineDateOfActivitySaturday: boolean
  onlineDateOfActivitySunday: boolean
  onlineDateOfActivityDetail: string
  admissionFeePerYear: number
  numberOfMembers: number
  isClubActivities: boolean
  appealingPoint1: string
  appealingPoint2: string
  appealingPoint3: string
  publicEmail: string
  twitterUrl: string
  facebookUrl: string
  instagramUrl: string
  lineUrl: string
  youtubeUrl: string
  homepageUrl: string
  peingUrl: string
  githubUrl: string
  tiktokUrl: string
  participationUrl: string
  mainImageUrl: string
  handbillImageUrl: string
  activityImageUrl1: string
  activityImageUrl2: string
  activityImageUrl3: string
  activityImageUrl4: string
  activityImageUrl5: string
  activityImageUrl6: string
  /** WordPressのURL */
  wpUrl: string
  /** WordPressのTag */
  wpTagTaxonomy: string
  /** WordPressの投稿を表示するかどうか */
  isViewWpPost: boolean
  createdAt: string
  updatedAt: string
}

export const isCircle = (v: any): v is Circle => v && v.type === 'Circle'